requirements.txt
mss>=9.0.0
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
pillow>=9.0.0
pynput>=1.7.0
pyautogui>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
